fix(firmware,sdk,web): code review security and correctness fixes
Firmware:
- safety: call motor_stop_all_isr() directly from e-stop ISR to eliminate
  20ms window where motors could run after e-stop assertion (F1)
- motor_control: add IRAM_ATTR motor_stop_all_isr() for ISR-safe shutdown;
  remove pre-lock e-stop check to eliminate spinlock race condition (F1, F2)
- safety: fix watchdog timeout comparison from > to >= to close off-by-one
  boundary where control loop stall at exact timeout would not trigger (F3)
- ota_signing: replace strncmp prefix check with full constant-time comparison
  of 'Bearer <token>' to eliminate timing side-channel attack (F4)
- encoder: use float division for raw CPR validation to avoid integer truncation (F7)
- robot_params: increase pi precision from 8 to 14 significant digits in
  MAX_LINEAR_VEL macro (F8)

Python SDK:
- vla_runner: torch.load weights_only=False -> True to prevent arbitrary code
  execution when loading model checkpoints (P1)
- text_encoder: replace MD5 with SHA-256 for token hashing (P2)
- vlm_runner: guard image.shape[2] with ndim check to handle grayscale/RGBA;
  separate JSONDecodeError from bare except to surface real errors (P3, P4)
- _ros2_node: warn instead of silently no-op when cmd_vel publisher not ready;
  warn on unknown image/depth encodings rather than dropping frames silently (P5, P6)
- payload_protocol: use strict ASCII decode and return None on UnicodeDecodeError
  instead of silently corrupting payload IDs with replacement chars (P7)
- async_client: replace assert statements with explicit RuntimeError so
  failures are not silently swallowed under python -O (P8)
- vla_runner: log warning when model returns non-tensor output instead of
  silently returning zero action vector with wrong shape (P9)
- exceptions: add ConnectionError/TimeoutError aliases for backwards compat
- provision_keys: raise ImportError instead of sys.exit(1) in library function
- gym/envs: wrap long docstrings to stay within 100-char line limit

TypeScript:
- payload-panel: add RailId type and VALID_RAILS runtime guard on toggleRail
  to prevent arbitrary strings being sent to the robot power service (T2)
- sensor-radar: use nullish coalescing on sensors[direction] lookup to handle
  unknown direction keys without crashing (T5)
- connection: emit console.warn with 'Unencrypted ws://' before blocking
  insecure non-localhost connections (aligns code with existing test contract)

+++ firmware/esp32/main/encoder.c
@@ -85,7 +85,7 @@ esp_err_t encoder_init(void)
 
 static void encoder_validate_config(void)
 {
-    uint16_t raw_cpr = ENCODER_CPR / GEAR_RATIO;
+    uint16_t raw_cpr = (uint16_t)((float)ENCODER_CPR / (float)GEAR_RATIO);
     if (raw_cpr < MIN_RAW_CPR) {
         ESP_LOGW(TAG, "Raw encoder CPR %d < minimum %d — high-speed PID may oscillate",
                  raw_cpr, MIN_RAW_CPR);
